Event commenting - on the API plus some tweaks
Event commenting fixed for the API.  Does not require authentication, so that
anonymous comments are possible via the API.  If user credentials are supplied
then they are checked accordingly and an error given if they are invalid.  If
not then this is an anonymous comment.  API docs are up to date.  The event
view page did not handle nameless commenters, this now shows the anonymous as
the talk comments list does.