Brings import up to date and inline with tz changes
Removes "update" functionality in import, changes field list to
only expect those that are actually used by the import.  Adds
support for types and tracks during the import.  Changes timezone
handling to bring in line with the rest of the application.  Docs
and validation updated to document and support the new format.