Fix for #GH-161, user/getdetail empty with numeric id
We were passing a SimpleXMLElement object in to the user_model getUserDetail
but it expects either a string or a number. The query works fine, since it
can __toString() itself but the is_numeric() check was failing.  Added a
sprintf in the wsaction so that at least we were passing valid data types to
the model