Replace Defensio with Akismet.
The Spamcheckservice class is from joindin-api with a change to the constructor to make it work with CI's load->library() system. We use Defensio in three places, so update all three to use Spamcheckservice and tidy the if() statements where they were testing against a string to make it clearer what's actually going on.